Addressing commments
